+++ cnxdb/config.py
@@ -24,13 +24,20 @@ def discover_settings(settings=None):
     if settings is None:
         settings = {}
 
+    # Retrieve the required url setting first, so that it can later be used
+    # as the default for the optional urls.
     common_url = os.environ.get('DB_URL', None)
+    if common_url is None:
+        if settings.get('db.common.url') is None:
+            raise RuntimeError("'DB_URL' environment variable "
+                               "OR the 'db.common.url' setting MUST "
+                               "be defined")
+        else:
+            common_url = settings['db.common.url']
+
+    # Set remaining urls, defaulting to the common url when not present.
     readonly_url = os.environ.get('DB_READONLY_URL', common_url)
     super_url = os.environ.get('DB_SUPER_URL', common_url)
-
-    if common_url is None and settings.get('db.common.url') is None:
-        raise RuntimeError("'DB_URL' environment variable "
-                           "OR the 'db.common.url' setting MUST be defined")
 
     new_settings = {
         'db.common.url': common_url,
@@ -38,6 +45,7 @@ def discover_settings(settings=None):
         'db.super.url': super_url,
     }
 
+    # Only amend the existing settings by setting the default.
     for k, v in new_settings.items():
         settings.setdefault(k, v)
     return settings